refactor: file name change and use relevant path
